feat: gate mint button refactored
