+++ apps/website/components/templates/gate-view/gate-view.tsx
@@ -1,23 +1,16 @@
 import dynamic from 'next/dynamic';
 import { useState, ComponentType } from 'react';
 
-import { useQuery } from '@tanstack/react-query';
 import { PartialDeep } from 'type-fest';
 
 import { Grid, Divider } from '@mui/material';
 
+import { useCredentialByGateId } from '../../../hooks/use-credential-by-gate-id';
 import { useGateStatus } from '../../../hooks/use-gate-status';
-import { useAuth } from '../../../providers/auth';
 import { Gates } from '../../../services/hasura/types';
-import { MintDialogProps } from '../../molecules/mint-dialog';
 import type { Props as HolderDialogProps } from '../../organisms/holder-dialog';
 import { GateViewSidebar } from './gate-view-sidebar';
 import { GateViewTasks } from './gate-view-tasks';
-
-const MintDialog: ComponentType<MintDialogProps> = dynamic(
-  () => import('../../molecules/mint-dialog').then((mod) => mod.MintDialog),
-  { ssr: false }
-);
 
 const HolderDialog: ComponentType<HolderDialogProps> = dynamic(
   () => import('../../organisms/holder-dialog').then((mod) => mod.HolderDialog),
@@ -30,28 +23,8 @@ type GateViewProps = {
 
 export function GateViewTemplate({ gateProps }: GateViewProps) {
   const [isHolderDialog, setIsHolderDialog] = useState(false);
-  const [isMintDialog, setMintModal] = useState(false);
-  const { me, gqlAuthMethods } = useAuth();
   const gateStatus = useGateStatus(gateProps);
-
-  const credential_id = me?.credentials?.find(
-    (cred) => cred?.gate_id === gateProps?.id
-  )?.id;
-
-  const { data: credential } = useQuery(
-    ['credential', credential_id],
-    () =>
-      gqlAuthMethods.credential({
-        id: credential_id,
-      }),
-    {
-      enabled: !!credential_id,
-    }
-  );
-
-  const isLimitExceeded = gateProps?.claim_limit
-    ? gateProps?.claim_limit <= gateProps?.holder_count
-    : false;
+  const credential = useCredentialByGateId({ gateId: gateProps?.id });
 
   return (
     <Grid
@@ -68,17 +41,10 @@ export function GateViewTemplate({ gateProps }: GateViewProps) {
           credentialId: gateProps?.id,
         }}
       />
-      <MintDialog
-        credential={credential?.credentials_by_pk}
-        isOpen={isMintDialog}
-        setOpen={setMintModal}
-      />
       <GateViewSidebar
         completedGate={gateStatus.isCompleted}
         credential={credential}
         gateProps={gateProps}
-        isLimitExceeded={isLimitExceeded}
-        setMintModal={setMintModal}
       />
       <Divider orientation="vertical" flexItem />
       <GateViewTasks
