+++ apps/website/components/templates/loyalty-program/components/CredentialListItem.tsx
@@ -8,11 +8,11 @@ import { brandColors } from '@gateway/theme';
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
 import { Chip, Stack, Typography, alpha } from '@mui/material';
 
-import SuccessfullyIcon from '../../../../../components/atoms/icons/successfully-icon';
-import SuccessfullyRoundedIcon from '../../../../../components/atoms/icons/successfully-rounded';
-import Loading from '../../../../../components/atoms/loading';
-import { ROUTES } from '../../../../../constants/routes';
-import { Gates } from '../../../../../services/hasura/types';
+import { ROUTES } from '../../../../constants/routes';
+import { Gates } from '../../../../services/hasura/types';
+import SuccessfullyIcon from '../../../atoms/icons/successfully-icon';
+import SuccessfullyRoundedIcon from '../../../atoms/icons/successfully-rounded';
+import Loading from '../../../atoms/loading';
 
 type Props = {
   gate: PartialDeep<Gates>;
