+++ apps/website/components/templates/loyalty-program/components/CredentialsList.tsx
@@ -6,14 +6,14 @@ import { TOKENS, brandColors } from '@gateway/theme';
 
 import { Stack, Typography, alpha } from '@mui/material';
 
-import { AlertCustom } from '../../../../../components/atoms/alert';
-import { useLoyaltyGatesCompleted } from '../../../../../hooks/use-loyalty-gates-completed';
-import { useAuth } from '../../../../../providers/auth';
+import { AlertCustom } from '../../../../components/atoms/alert';
+import { useLoyaltyGatesCompleted } from '../../../../hooks/use-loyalty-gates-completed';
+import { useAuth } from '../../../../providers/auth';
 import {
   Gate_Progress,
   Gates,
   Loyalty_Program,
-} from '../../../../../services/hasura/types';
+} from '../../../../services/hasura/types';
 import { CredentialListItem } from './CredentialListItem';
 
 type Props = {
