+++ apps/website/locales/en/credential.json
@@ -30,5 +30,9 @@
     "issuer-id": "Issuer ID",
     "recipient-id": "Recipient ID",
     "issuance-date": "Issuance date"
+  },
+  "mint": {
+    "mint_nft": "Mint as NFT",
+    "verify_mint_transaction": "Verify mint transaction"
   }
 }
