+++ apps/website/locales/en/loyalty-program.json
@@ -12,7 +12,6 @@
   },
   "sidebar": {
     "display-credential": "See credential",
-    "mint-nft": "Mint as NFT",
     "data-model-id": "Data model ID"
   }
 }
